feat(install): add install command
This commit adds a new `install` command to the `ddns` application. The `install` command is responsible for installing the necessary systemd service and timer files for the `ddns` application.

The `install` command creates the systemd service file `ddns.service` and the systemd timer file `ddns.timer` in the user's config directory. The contents of these files are read from the `docs/ddns/ddns.service` and `docs/ddns/ddns.timer` files respectively.

After creating the files, the `install` command enables and starts the `ddns.timer` using the `systemctl --user enable --now ddns.timer` command.

This new command allows users to easily install and configure the `ddns` application as a systemd service.