feat(cf/tunnel/list): add `cf tunnel get` command (#29)
* fix(api): refactor Cloudflare client initialization

Refactor the Cloudflare client initialization in the `mod.rs` file under the `api/cloudflare` directory. The changes include:
- Replacing the usage of `reqwest::Client` with `ClientBuilder` for more flexibility in setting default headers.
- Adding the `api` field to the `Cloudflare` struct to allow for custom API endpoints.
- Modifying the `new` function in the `Cloudflare` implementation to accept an optional custom API endpoint.
- Updating the initialization of the `reqwest::Client` with default headers and custom API endpoint.
- Removing the `zone` field from the `Client` struct as it is no longer needed.

These changes aim to improve the flexibility and customization options for the Cloudflare client initialization process.

* fix(ddns/dns/install): update ddns service executable path

Update the executable path for the ddns service in the systemd unit file to include the 'dns' subcommand. This change ensures that the correct command is executed when starting the ddns service.

* fix(cf/dns/install): update Taskfile.yaml and move ddns related files to cf directory

- Update TARGETS in Taskfile.yaml to include 'cf' instead of 'ddns'
- Move ddns related files from 'docs/ddns' to 'docs/cf'
- Move ddns related files from 'src/bin/ddns' to 'src/bin/cf'
- Update paths in ddns install.rs to reflect the changes

The changes are made to align the project structure and naming conventions with the use of 'cf' instead of 'ddns' for Cloudflare related functionalities.

* fix(cf): update token retrieval endpoint

In `src/bin/cf/cmd/mod.rs`, the endpoint for retrieving the Cloudflare token has been updated from "CLOUDFLARE_TOKEN" to "cloudflare.com". This change ensures that the correct endpoint is used for retrieving the Cloudflare token.

* feat(cf/tunnel/get): add functionality to retrieve and display configurations for Cloudflare Tunnel

- Added a new module `tunnel` under `cmd` in the Cloudflare CLI tool.
- Implemented a new subcommand `get` to retrieve and display configurations for Cloudflare Tunnel.
- Created a new struct `Cmd` to handle the `get` subcommand.
- Implemented the `run` method for the `get` subcommand to fetch and display the configurations.
- Utilized the `AccountsArgsFromGlobal` struct to handle global arguments for Cloudflare API.
- Modified the table formatting to display the retrieved configurations with proper alignment and colors.

This commit introduces the ability to fetch and visualize Cloudflare Tunnel configurations in the Cloudflare CLI tool.