feat(cf/tunnel/put): add command `cf tunnel put` (#35)
* feat(cf/tunnel/put): add new configurations for Cloudflare Tunnel

Added new configurations for Cloudflare Tunnel in the `put` endpoint. The configurations include
ingress settings for three new services: `alist`, `glances`, and `jellyfin`. Each service has a
specific hostname and service URL. The `alist` service is available at `http://127.0.0.1:5244`,
the `glances` service at `http://127.0.0.1:61208`, and the `jellyfin` service at `http://127.0.0.1:8096`.

These new configurations enhance the functionality of the Cloudflare Tunnel by allowing traffic
routing to these services.

* fix(cf/tunnel/put: update tunnel and DNS configurations

- Updated the `cspell.json` file to include the `cfargotunnel` word.
- Added `Clone` trait to the `ZonesArgs` and `ZonesArgsFromGlobal` structs.
- Modified the `ZonesArgsFromGlobal` struct to have public fields.
- Changed the visibility of the `zones` method in `ZonesArgsFromGlobal` to public.
- Updated the `update` method in the `Cmd` struct to use `try_join` for better error handling.
- Refactored the `update_tunnel` and `update_dns` functions for updating tunnel and DNS configurations.
- Added a new `ZonesArgs` struct for handling zone and name parameters.
- Implemented async functions `update_tunnel` and `update_dns` to update tunnel and DNS configurations respectively.

The changes were made to improve the handling and updating of tunnel and DNS configurations in the Cloudflare tunneling system.

* fix(gfw/sub/conf): add optional url field to URLTest model

Import the `Optional` type from the `typing` module to allow the `url` field in the `URLTest` model to be nullable. Set a default value of "https://cp.cloudflare.com" for the `url` field. This change enhances the flexibility of the `URLTest` model by making the `url` field optional.